refactor: move MergeKeyStore into core
