+++ packages/core/src/key_store/index.ts
@@ -1,2 +1,3 @@
 export { InMemoryKeyStore } from './in_memory_key_store';
 export { KeyStore } from './keystore';
+export { MergeKeyStore } from './merge_key_store';
